# openapi.yaml
openapi: "3.1.0"
info:
  title: "Aegis API"
  version: "1.0.0"
  description: "Orchestrate Claude Code sessions via REST API. Aegis wraps Claude Code in ACP sessions\nand exposes everything through a unified API.\n\n**Base URL:** `http://localhost:9100`\n\n**Authentication:** All endpoints except `/v1/health` and `/health` require a bearer token.\nPass `Authorization: Bearer <your-token>` header, or set `AEGIS_AUTH_TOKEN` env var.\n\n**Session lifecycle:** `working` → `idle` → `permission_prompt` → `asking` → `stalled`\n\n**API versioning:** All `/v1/` responses include `X-Aegis-API-Version: 1`.\nLegacy unversioned paths (e.g. `/health`, `/sessions`) emit `Deprecation: true`,\n`Sunset`, and `X-API-Deprecated` headers. See `docs/api-versioning.md` for the\nfull versioning policy.\n"
servers:
  - url: "http://localhost:9100"
    description: "Local Aegis server"
security:
  - BearerAuth: []
paths:
...
  "/v1/sessions/{id}":
    get:
      operationId: "getSession"
      summary: "Get session details"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Session details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"



    delete:
      operationId: "killSession"
      summary: "Kill a session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "204":
          description: "Session killed"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  "/v1/sessions/{id}/pause":
    post:
      summary: "Pause session"
      description: "Temporarily pause a running session."
      operationId: "postV1SessionsIdPause"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                reason:
                  type: "string"
                  minLength: 1
                  maxLength: 2048

                requestedBy:
                  type: "string"
                  minLength: 1
                  maxLength: 256

                idempotencyKey:
                  type: "string"
                  minLength: 1
                  maxLength: 256

                metadata:
                  type: "object"
                  propertyNames:
                    type: "string"

                  additionalProperties:
                    anyOf:
                      - type: "string"

                      - type: "number"

                      - type: "boolean"

                      - type: "null"





              required:
                - "reason"

              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"

                  pausedAt:
                    type: "number"


                required:
                  - "ok"
                  - "pausedAt"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "404":
          description: "Not found"

        "409":
          description: "Session not in a pausable state"
  "/v1/sessions/{id}/resume":
    post:
      summary: "Resume session"
      description: "Resume a paused session."
      operationId: "postV1SessionsIdResume"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                resumedBy:
                  type: "string"
                  minLength: 1
                  maxLength: 256

                resumeMetadata:
                  type: "object"
                  propertyNames:
                    type: "string"

                  additionalProperties:
                    anyOf:
                      - type: "string"

                      - type: "number"

                      - type: "boolean"

                      - type: "null"





              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"

                  resumedAt:
                    type: "number"


                required:
                  - "ok"
                  - "resumedAt"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "404":
          description: "Not found"

        "409":
          description: "Session not in a resumable state"
  "/v1/sessions/{id}/cancel":
    post:
      summary: "Cancel session"
      description: "Cancel the current turn in a session."
      operationId: "postV1SessionsIdCancel"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                force:
                  type: "boolean"


              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"

                  cancelledAt:
                    type: "number"


                required:
                  - "ok"
                  - "cancelledAt"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "404":
          description: "Not found"

        "409":
          description: "Session not in a cancellable state"
  "/v1/sessions/{id}/intervention/start":
    post:
      summary: "Start human intervention"
      description: "Flag a session as under human intervention."
      operationId: "postV1SessionsIdInterventionStart"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                interventionBy:
                  type: "string"
                  minLength: 1
                  maxLength: 256


              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"

                  interventionStartedAt:
                    type: "number"


                required:
                  - "ok"
                  - "interventionStartedAt"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "404":
          description: "Not found"

        "409":
          description: "Intervention already active"
  "/v1/sessions/{id}/intervention/complete":
    post:
      summary: "Complete human intervention"
      description: "Resolve an active human intervention on a session."
      operationId: "postV1SessionsIdInterventionComplete"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                completedBy:
                  type: "string"
                  minLength: 1
                  maxLength: 256

                guidance:
                  type: "string"
                  maxLength: 8192


              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"

                  interventionCompletedAt:
                    type: "number"


                required:
                  - "ok"
                  - "interventionCompletedAt"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "404":
          description: "Not found"

        "409":
          description: "No active intervention"
  "/v1/sessions/{id}/intervention":
    get:
      summary: "Get intervention status"
      description: "Return the current intervention state for a session."
      operationId: "getV1SessionsIdIntervention"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  active:
                    type: "boolean"

                  startedAt:
                    anyOf:
                      - type: "number"

                      - type: "null"



                  startedBy:
                    anyOf:
                      - type: "string"

                      - type: "null"



                  guidance:
                    anyOf:
                      - type: "string"

                      - type: "null"




                required:
                  - "active"
                  - "startedAt"
                  - "startedBy"
                  - "guidance"

                additionalProperties: false




        "404":
          description: "Not found"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "API key"
  parameters:
    sessionId:
      name: "id"
      in: "path"
      required: true
      schema:
        type: "string"

      description: "Session ID"
...
  responses:
...
    Unauthorized:
      description: "Missing or invalid authentication token"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Unauthorized"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Not found"
  schemas:
...
    SessionInfo:
      type: "object"
      properties:
        id:
          type: "string"

        windowId:
          type: "string"

        windowName:
          type: "string"

        workDir:
          type: "string"

        claudeSessionId:
          type: "string"

        jsonlPath:
          type: "string"

        byteOffset:
          type: "integer"

        monitorOffset:
          type: "integer"

        status:
          type: "string"
          enum:
            - "idle"
            - "working"
            - "compacting"
            - "context_warning"
            - "waiting_for_input"
            - "permission_prompt"
            - "plan_mode"
            - "ask_question"
            - "bash_approval"
            - "settings"
            - "error"
            - "rate_limit"
            - "unknown"


        createdAt:
          type: "integer"

        lastActivity:
          type: "integer"

        stallThresholdMs:
          type: "integer"

        permissionStallMs:
          type: "integer"

        permissionMode:
          type: "string"

        settingsPatched:
          type: "boolean"

        hookSettingsFile:
          type: "string"

        hookSecret:
          type: "string"

        lastHookAt:
          type: "integer"

        activeSubagents:
          type: "array"
          items:
            type: "string"


        permissionPromptAt:
          type: "integer"

        permissionRespondedAt:
          type: "integer"

        lastHookReceivedAt:
          type: "integer"

        lastHookEventAt:
          type: "integer"

        model:
          type: "string"

        lastDeadAt:
          type: "integer"

        ccPid:
          type: "integer"

        parentId:
          type: "string"

        children:
          type: "array"
          items:
            type: "string"


        permissionPolicy:
          type: "array"
          items:
            type: "object"


        permissionProfile:
          type: "object"

        prd:
          type: "string"

        ownerKeyId:
          type: "string"
...